requirements.txt
django==4.0.1
djangorestframework==3.13.1
django-cors-headers==3.11.0
requests==2.27.1
python-dotenv==0.19.2
textblob==0.17.1
numpy>=1.26.0
pandas==2.0.2
scikit-learn==1.5.2
kaggle==1.5.16
pytz==2023.3
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Azure deployment requirements
gunicorn==21.2.0
whitenoise==6.6.0
mysqlclient==2.2.4  # MySQL client library
django-storages[azure]==1.14.2  # For Azure Storage
azure-storage-blob==12.19.0  # For Azure Blob Storage
python-decouple==3.8  # For environment variables

#pip install -r requirements.txt
